Micro-Functionalities Added {Active-Pending status circle color-based & report-tooltip}
Status circle color-based on status of project added to indicate to the user or admin which project is active and which on pending.

Report tooltip added to warn user that if he submits another report in the specific project it will overwrite any previous report of his at the specific project's id.